+++ HISTORY.rst
@@ -18,8 +18,9 @@ picows Release History
 * User on_ws_connect and on_ws_frame implementation can now signalize protocol errors by raising WSProtocolError
 * Add is_disconnected property to WSTransport.
 * Fix send_* methods raising exceptions when attempting to send after connection abort and without prior CLOSE frame.
-* Add missing body attribute in WSUpgradeResponse at the client side.
-* WSTransport.request was missing for the server side.
+* Add missing WSUpgradeResponse.body attribute at the client side.
+* Add missing WSTransport.request attribute for the server side.
+* ws_connect can now accept listener_factory that takes WSUpgradeRequest, WSUpgradeResponse as arguments. Old argument-less client_factory also works.
 
 1.19.0 (2026-04-24)
 ------------------